requirements.txt
Flask==0.10.1
Django==1.5.2
django-tastypie==0.10.0
django-social-auth==0.7.26
django-registration==1.0
pymongo==2.6.3
boilerpipe==1.2.0.0
psycopg2==2.5.1
mimeparse==0.1.3
PIL==1.1.7
south==0.8.2
yolk==0.4.3
